Read the raw tri-axial accelerometry data csv file by ActiGraph GT3X+.

ReadGT3XPlus reads the accelerometry data collected by ActiGraph GT3XPlus in csv files generated by ActiLife software. It automatically parses the header of the csv file to acquire the setting of the device.

ReadGT3XPlus(filename)

Arguments

filename
The name of the csv file.

Value

The ReadGT3XPlus returns an object of class "GT3XPlus". This class of object is supported by functions computeActivityIndex. A object of class "GT3XPlus" is a list containing at least the following components:

SN: Serial Number of the accelerometer

StartTime: Start time of the data collection

StartDate: Start date of the data collection

Epoch: The Epoch time of each observation. If sample rate Hertz>1, then Epoch=00:00:00

DownloadTime: Download time of the data

DownloadDate: Download date the data

Hertz: Sampling Rate

Raw: a data frame with 5 columns containing the date, time and acceleration in X, Y and Z axes for each observation.

Details

The function is tested on the csv files generated by ActiLife6, which have exactly 10 lines of headers, containing information about the device name, the starting and ending date/time of data collection, the sample rate, and the downloading date/time, etc. The 11th may further be omitted, if it is the header for the tri-axial acceleration time series. The function only reads the first 3 columns from then, if more are present.

Examples

filename = system.file("extdata","sample_GT3X+.csv.gz",package="ActivityIndex")
res = ReadGT3XPlus(filename)